requirements.txt
python-binance
pandas
requests
python-dotenv
flask
